+++ src/command/react.rs
@@ -1,10 +1,6 @@
 //! The react command converts an ASCII-alphanumeric string into a series of
 //! reaction emojis, which it adds to a target message.
 
-use serenity::model::{
-    channel::ReactionType,
-    id::{ChannelId, MessageId},
-};
 use tracing::{instrument, warn};
 
 use std::collections::HashMap;
@@ -12,20 +8,15 @@ use std::collections::HashMap;
 use super::{CommandError, Response};
 
 #[instrument]
-pub fn react(
-    channel_id: ChannelId,
-    command_id: MessageId,
-    target_id: MessageId,
-    raw_reaction: String,
-) -> Result<Vec<Response>, CommandError> {
+pub fn react(input: String) -> Result<Response, CommandError> {
     // Ignore spaces by removing them before checking if the input is valid.
-    let raw_reaction = raw_reaction.replace(" ", "");
+    let input = input.replace(" ", "");
 
     // An input string is only valid if it is entirely composed of alphanumeric
     // characters, and if each one only appears once.
-    let non_alphanum = raw_reaction.contains(|c: char| !c.is_alphanumeric());
+    let non_alphanum = input.contains(|c: char| !c.is_alphanumeric());
     let valid = !non_alphanum
-        && raw_reaction
+        && input
             .chars()
             .fold(HashMap::new(), |mut acc, next| {
                 *acc.entry(next.to_ascii_uppercase()).or_insert(0) += 1;
@@ -35,46 +26,30 @@ pub fn react(
             .all(|&v| v == 1);
 
     if valid {
-        let mut responses = vec![Response::DeleteMessage {
-            channel_id,
-            message_id: command_id,
-        }];
+        let response = Response::React {
+            reactions: to_reactions(&input),
+        };
 
-        let mut reactions = to_reactions(&raw_reaction)
-            .into_iter()
-            .map(|reaction| Response::React {
-                channel_id,
-                message_id: target_id,
-                reaction,
-            })
-            .collect::<Vec<_>>();
-
-        responses.append(&mut reactions);
-
-        Ok(responses)
+        Ok(response)
     } else if non_alphanum {
         warn!("string contains non-alphanumeric characters");
 
-        Err(CommandError::NonAlphanumeric {
-            original: raw_reaction,
-        })
+        Err(CommandError::NonAlphanumeric { original: input })
     } else {
         warn!("string contains repeated characters");
 
-        Err(CommandError::Repetition {
-            original: raw_reaction,
-        })
+        Err(CommandError::Repetition { original: input })
     }
 }
 
 const VARIATION_SELECTOR_16: u32 = 0xfe0f;
 const COMBINING_ENCLOSING_KEYCAP: u32 = 0x20e3;
 
-/// Convert a string to a sequence of reactions representing its characters,
-/// using regional indicators for alphabetic characters and keycap sequences for
+/// Convert a string to a sequence of emojis representing its characters, using
+/// regional indicators for alphabetic characters and keycap sequences for
 /// numerals. Any non-ascii-alphanumeric characters are simply left as-is in the
 /// output string.
-fn to_reactions(input: &str) -> Vec<ReactionType> {
+fn to_reactions(input: &str) -> Vec<String> {
     input
         .chars()
         .map(|c| match c {
@@ -95,6 +70,5 @@ fn to_reactions(input: &str) -> Vec<ReactionType> {
             }
             _ => c.to_string(),
         })
-        .map(ReactionType::Unicode)
         .collect()
 }