+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Miscellaneous
 
+- Made command processing independent of any Discord context (#9).
 - Split crate into a library and a binary, so that hatysa can be used from other Rust crates.
 
 ## 0.2.1